+++ scantpaper/savethread.py
@@ -493,7 +493,9 @@ class SaveThread(Importhread):
                     resolution=page.resolution,
                     text_layer=page.text_layer,
                 )
-                row = self.replace_page(new, self.find_page_number_by_page_id(page.id))
+                row = self.replace_page(
+                    new, self.find_page_number_by_initial_id(page.id), page.id
+                )
                 request.data(
                     {
                         "type": "page",