+++ changelog.md
@@ -1,6 +1,9 @@
 ## 3.0.3 (2026-04-03)
+
 * + `Error during device I/O` to errors triggering "Error opening the last device used" dialog
 * Question caching device list including libusb devices
+* Fix race conditions when new actions triggered before previous finished
+  Closes #55 (Save error)
 
 
 ## 3.0.2 (2026-03-27)